Remove ark-circom
