+++ app/crates/core/witness/Cargo.toml
@@ -5,37 +5,22 @@ edition.workspace = true
 repository.workspace = true
 license.workspace = true
 publish.workspace = true
-description = "Browser WASM module for Circom witness generation using ark-circom"
+description = "Circom witness generation via circom-witness-rs operation graphs"
 
 [dependencies]
 anyhow = { workspace = true }
 
-# Arkworks
-ark-bn254 = { workspace = true }
-ark-circom = { workspace = true }
-
 # Graph-based witness generation
 circom-witness-rs = { workspace = true }
 ruint = { workspace = true }
-# circom-witness-rs's black box functions use arkworks 0.5 `Fr` (workspace uses workspace 0.6 arkworks)
+# circom-witness-rs's black box functions use arkworks 0.5 `Fr`
 ark-bn254-05 = { package = "ark-bn254", version = "0.5", default-features = false }
 ark-ff-05 = { package = "ark-ff", version = "0.5", default-features = false }
 
-# BigInt support for input parsing
-num-bigint = { workspace = true }
-
 # JSON parsing
 serde_json = { workspace = true, default-features = false, features = ["alloc", "std"] }
-
-[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-ark-circom = { workspace = true, features = ["parallel"] }
-# see https://github.com/NethermindEth/stellar-private-payments/issues/192
-wasmer = { workspace = true, features = ["cranelift"] }
-
-[target.'cfg(target_arch = "wasm32")'.dependencies]
-ark-circom = { workspace = true }
-wasmer = { workspace = true, features = ["js", "js-default"] }
 # circom-witness-rs pulls rand -> getrandom 0.2, which needs the "js" feature for web
+[target.'cfg(target_arch = "wasm32")'.dependencies]
 getrandom = { workspace = true, features = ["js"] }
 
 [lints]
